Fix duplicated rows
Querying with 'gte' will lead to duplicated rows if the last
oplog entry is for a collection that is supposed to be replicated.
Because of the change done in efd5e0fd769b749f247266a88f617a0e617d380a
this will not happen for every sync as it's quite likely that the last
oplog entry references a modification in a different collection.

(cherry picked from commit afbf251931a75abd118d01c96a00b94fecd227d2)